Moved relevant query helpers to live with redirects instead of utils folder
