+++ src/components/PayloadRedirects/queries.ts
@@ -1,8 +1,8 @@
-import type { Config } from "src/payload-types"
+import type { Config } from "@/payload-types"
 
 import configPromise from "@payload-config"
-import { getPayload } from "payload"
 import { unstable_cache } from "next/cache"
+import { getPayload } from "payload"
 
 type Collection = keyof Config["collections"]
 
@@ -22,6 +22,19 @@ async function getDocument(collection: Collection, slug: string, depth = 0) {
   return page.docs[0]
 }
 
+async function getRedirects(depth = 1) {
+  const payload = await getPayload({ config: configPromise })
+
+  const { docs: redirects } = await payload.find({
+    collection: "redirects",
+    depth,
+    limit: 0,
+    pagination: false,
+  })
+
+  return redirects
+}
+
 /**
  * Returns a unstable_cache function mapped with the cache tag for the slug
  */
@@ -29,4 +42,15 @@ async function getDocument(collection: Collection, slug: string, depth = 0) {
 export const getCachedDocument = (collection: Collection, slug: string) =>
   unstable_cache(async () => getDocument(collection, slug), [collection, slug], {
     tags: [`${collection}_${slug}`],
+  })
+
+/**
+ * Returns a unstable_cache function mapped with the cache tag for 'redirects'.
+ *
+ * Cache all redirects together to avoid multiple fetches.
+ */
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export const getCachedRedirects = () =>
+  unstable_cache(async () => getRedirects(), ["redirects"], {
+    tags: ["redirects"],
   })
